add categorie

+++ app_client/AdminLTE/admin/gestionusers/gestionusers.controller.js
@@ -10,7 +10,7 @@
         $scope.names = ["Emil", "Tobias", "Linus"];
         $scope.message = "hello ahmed";
 
-
+        $scope.users =[];
         $http({
             method: 'GET',
             url: '/gestionusers'
@@ -22,45 +22,58 @@
 
 
 
+        $scope.showSearch =function showSearch(value) {
+            if(value==""){
+                $http({
+                    method: 'GET',
+                    url: '/gestionusers'
+                }).success(function (data) {
+                    $scope.users = data; // response data
+                }).error(function (response) {
+                    console.log('error message :',response);
+                });
+            }else{
+                $http({
+                    method: 'GET',
+                    url: 'gestionusers/usersearch/'+value
+                }).success(function (data) {
+                    $scope.users = data; // response data
+                }).error(function (response) {
+                    console.log('error message :',response);
+                });
+            }
+        }
+
 
 
 
 
 
         $scope.deleteUser = function deleteUser(user) {
+            swal({
+                    title : '',
+                    text: "Êtes-vous sûr de vouloir supprimer cet utilisateur ?",
+                    type: "warning",
+                    showCancelButton: true,
+                    cancelButtonText: "Annuler",
+                    closeOnConfirm: false,
+                    confirmButtonText: "Confirmer",
+                    confirmButtonColor: "#ec6c62"
+                }, function() {
 
-            $scope.messaguser = "delete success";
-            $http.delete("/gestionusers/" + user._id).success(function (response) {
-                var index = $scope.users.indexOf(user)
-                $scope.users.splice( index ,1);
-                $scope.ALertSuccess(user.name +' supprimer avec success');
-                //   $scope.users.pop(user);
-            })
+                $http.delete("/gestionusers/" + user._id).success(function (response) {
+                    var index = $scope.users.indexOf(user)
+                    $scope.users.splice(index, 1);
+                    swal("", "utilisateur a été supprimé avec succès!", "success");
+                }).error(function (response) {
+                    swal("Oops", "We couldn't connect to the server!", "error");
+                });
 
+            });
         }
 
-        $scope.ALertSuccess =function ALertSuccess(message) {
-            $scope.msgalert=message;
-            $scope.alert = true;
-        }
 
 
-
-
-        // ici pour pagination
-        $scope.curPage = 0;
-        $scope.pageSize = 2;
-
-        $scope.numberOfPages = function () {
-            return Math.ceil($scope.users.length / $scope.pageSize);
-        };
-
-        //ici pour serche
-        $scope.sort = function (keyname) {
-            $scope.sortKey = keyname;
-            $scope.reverse = !$scope.reverse;
-
-        }
 
 
     }
