+++ app_client/AdminLTE/admin/sservices/sservices.controller.js
@@ -20,12 +20,26 @@
         });
 
         $scope.deleteSservice = function deletesservice(service) {
-            $scope.messaguser = "delete success";
-          //  $scope.sservices.splice( service._id,3);
-            $http.delete("/gestionusers/sservice/" + service._id).success(function (response) {
-                var index = $scope.sservices.indexOf(service)
-              $scope.sservices.splice( index ,1);
-            })
+            swal({
+                title: "vous êtes sûr ?",
+                text: "Êtes-vous sûr de vouloir supprimer cette sous catégorie ?",
+                type: "warning",
+                showCancelButton: true,
+                closeOnConfirm: false,
+                confirmButtonText: "oui , Supprimer!",
+                confirmButtonColor: "#ec6c62"
+            }, function() {
+
+                $http.delete("/gestionusers/sservice/" + service._id).success(function (response) {
+                    var index = $scope.sservices.indexOf(service)
+                    $scope.sservices.splice( index ,1);
+                    swal("Supprimer!", "sous catégorie a été supprimé avec succès!", "success");
+                }).error(function (response) {
+                    swal("Oops", "We couldn't connect to the server!", "error");
+                });
+
+            });
+
         }
