+++ app_client/AdminLTE/admin/ville/ville.controller.js
@@ -21,11 +21,27 @@
 
 
         $scope.deleteVille = function deleteVille(ville) {
-            $scope.messaguser = "delete success";
-            $http.delete("/gestionusers/ville/" + ville._id).success(function (response) {
-                var index = $scope.villes.indexOf(ville)
-                $scope.villes.splice( index ,1);
-            })
+            swal({
+                title: "vous êtes sûr ?",
+                text: "Êtes-vous sûr de vouloir supprimer cette Ville ?",
+                type: "warning",
+                showCancelButton: true,
+                cancelButtonText: "Annuler",
+                closeOnConfirm: false,
+                confirmButtonText: "oui , Supprimer!",
+                confirmButtonColor: "#ec6c62"
+            }, function() {
+
+                $http.delete("/gestionusers/ville/" + ville._id).success(function (response) {
+                    var index = $scope.villes.indexOf(ville)
+                    $scope.villes.splice( index ,1);
+                    swal("Supprimer!", "ville a été supprimé avec succès!", "success");
+                }).error(function (response) {
+                    swal("Oops", "We couldn't connect to the server!", "error");
+                });
+
+            });
+
         }
